+++ packages/signals/pubspec.yaml
@@ -3,7 +3,7 @@ description: "Reactivity made simple. Do more by doing less. Supports Flutter an
 repository: https://github.com/rodydavis/signals.dart
 homepage: https://dartsignals.dev/
 documentation: https://dartsignals.dev/
-version: 6.0.0-rc.1
+version: 6.0.0-rc.2
 
 environment:
   sdk: ">=3.0.0 <4.0.0"
@@ -12,8 +12,8 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  signals_core: ^6.0.0-rc.1
-  signals_flutter: ^6.0.0-rc.1
+  signals_core: ^6.0.0-rc.2
+  signals_flutter: ^6.0.0-rc.2
 
 dev_dependencies:
   flutter_test:
