+++ packages/signals_core/CHANGELOG.md
@@ -2,8 +2,13 @@
 
 - Switching to **preact_signals** package for core implementation
 - Removing old deprecated methods
-- Add new mixins: EventSinkSignalMixin, SinkSignalMixin, StreamSignalMixin
+- Add new mixins: EventSinkSignalMixin, SinkSignalMixin, StreamSignalMixin, SetSignalMixin, ListSignalMixin, MapSignalMixin, QueueSignalMixin, ChangeStackSignalMixin, IterableSignalMixin
 - Update AsyncSignal to implement EventSink
+- Removing previous/initial value from Signal and Computed in favor of TrackedSignal, TrackedSignalMixin and trackedSignal()
+- Update SignalsObserver to include value for signal created (instead of peek())
+- Updated examples
+- Remove callback to signal
+- Remove toSignal extension method (causing unintended casts) in favor of .$ for Object/Object?
 
 ## 5.5.0
