+++ packages/signals_core/lib/signals_core.dart
@@ -1,11 +1,11 @@
 /// Core signals library
 library signals_core;
 
+export 'src/devtool.dart';
 export 'src/async/connect.dart';
 export 'src/extensions/future.dart';
 export 'src/extensions/stream.dart';
 export 'src/extensions/signal.dart';
-export 'src/extensions/callback.dart';
 export 'src/extensions/object.dart' hide OptionalSignalObjectUtils;
 export 'src/utils/timer.dart';
 export 'src/async/state.dart';
@@ -19,4 +19,11 @@ export 'src/core/signals.dart';
 export 'src/mixins/event_sink.dart';
 export 'src/mixins/sink.dart';
 export 'src/mixins/stream.dart';
-export 'src/devtool.dart';
+export 'src/mixins/changestack.dart';
+export 'src/mixins/iterable.dart';
+export 'src/mixins/list.dart';
+export 'src/mixins/set.dart';
+export 'src/mixins/map.dart';
+export 'src/mixins/queue.dart';
+export 'src/mixins/tracked.dart';
+
