+++ packages/signals_core/lib/src/extensions/stream.dart
@@ -19,40 +19,6 @@ extension SignalStreamUtils<T> on Stream<T> {
   /// final stream = createStream();
   /// final signal = stream.toSignal();
   /// ```
-  @Deprecated('use [toStreamSignal] instead')
-  StreamSignal<T> toSignal({
-    bool? cancelOnError,
-    String? debugLabel,
-    T? initialValue,
-    bool autoDispose = false,
-    bool lazy = true,
-    List<ReadonlySignal<dynamic>> dependencies = const [],
-    void Function()? onDone,
-  }) {
-    return toStreamSignal(
-      cancelOnError: cancelOnError,
-      debugLabel: debugLabel,
-      initialValue: initialValue,
-      autoDispose: autoDispose,
-      lazy: lazy,
-      dependencies: dependencies,
-      onDone: onDone,
-    );
-  }
-
-  /// Convert a stream to a signal
-  ///
-  /// ```dart
-  /// import 'package:signals/signals.dart';
-  ///
-  /// Stream<int> createStream() async* {
-  ///     yield 1;
-  ///     yield 2;
-  ///     yield 3;
-  /// }
-  /// final stream = createStream();
-  /// final signal = stream.toSignal();
-  /// ```
   ///
   /// For returning a signal with the value that can be accessed sync use
   /// stream.toSyncSignal instead.
