+++ packages/signals_core/lib/src/value/change_stack.dart
@@ -13,7 +13,7 @@ part of 'value.dart';
 /// s.redo();
 /// print(s.value); // 3
 /// ```
-class ChangeStackSignal<T> extends Signal<T> {
+class ChangeStackSignal<T> extends Signal<T> with ChangeStackSignalMixin<T> {
   /// Change stack signal that can be used to call undo/redo on a value.
   ///
   /// ```dart
@@ -29,83 +29,13 @@ class ChangeStackSignal<T> extends Signal<T> {
   /// ```
   ChangeStackSignal(
     super.value, {
-    this.limit,
+    int? limit,
     super.debugLabel,
     super.autoDispose,
-  });
-
-  /// Max values to keep in history
-  final int? limit;
-  final Queue<SignalChange<T>> _undo = ListQueue();
-  final Queue<SignalChange<T>> _redo = ListQueue();
-
-  /// List of changes in the history
-  Iterable<SignalChange<T>> get history => _undo;
-
-  /// List of changes in the redo stack
-  Iterable<SignalChange<T>> get redos => _redo;
-
-  /// Can redo the previous change
-  bool get canRedo => _redo.isNotEmpty;
-
-  /// Can undo the previous change
-  bool get canUndo => _undo.isNotEmpty;
-
-  @override
-  bool set(
-    T val, {
-    bool force = false,
   }) {
-    final prev = super.value;
-    final updated = super.set(val, force: force);
-    if (!updated) return false;
-    _undo.addLast((previousValue: prev, value: val));
-    _redo.clear();
-    if (limit != null && _undo.length > limit!) {
-      _undo.removeFirst();
-    }
-    return true;
-  }
-
-  /// Redo Previous Undo
-  void redo() {
-    if (!canRedo) return;
-    final change = _redo.removeFirst();
-    super.set(change.value, force: true);
-    _undo.addLast(change);
-  }
-
-  /// Undo Last Change
-  void undo() {
-    if (!canUndo) return;
-    final change = _undo.removeLast();
-    super.set(change.previousValue, force: true);
-    _redo.addFirst(change);
-  }
-
-  /// Clear the history for redo and undo
-  void clear() {
-    clearUndo();
-    clearRedo();
-  }
-
-  /// Clear undo stack
-  void clearUndo() {
-    _undo.clear();
-  }
-
-  /// Clear redo stack
-  void clearRedo() {
-    _redo.clear();
+    this.limit = limit;
   }
 }
-
-/// Signal change that contains a snapshot of the
-/// previous value and next value
-typedef SignalChange<T> = ({
-  T previousValue,
-  T value,
-});
 
 /// Change stack signal that can be used to call undo/redo on a value.
 ///
