+++ packages/signals_core/test/extensions/object_test.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: unnecessary_type_check
+
 import 'package:signals_core/extended.dart';
 import 'package:signals_core/signals_core.dart';
 import 'package:test/test.dart';
@@ -7,7 +9,7 @@ void main() {
   group('Object to signal extensions', () {
     test('Object', () {
       final val = Object();
-      final s = val.asSignal();
+      final s = val.$;
 
       expect(s.value, val);
     });
@@ -17,13 +19,15 @@ void main() {
       final s = val.$;
 
       expect(s.value, val);
+      expect(s is Signal<Object>, true);
     });
 
     test('Object?', () {
       final Object? val = null;
-      final s = val.toSignal();
+      final s = val.$;
 
       expect(s.value, null);
+      expect(s is Signal<Object?>, true);
     });
   });
 }
