+++ packages/signals_core/test/value/list_test.dart
@@ -12,14 +12,6 @@ void main() {
     }) create,
   ) {
     group('$message test', () {
-      test('initialValue/previousValue throw an error', () {
-        final List<String> list = <String>['a', 'b', 'c'];
-        final s = create(list, autoDispose: true);
-
-        expect(() => s.initialValue, throwsA(isA<Exception>()));
-        expect(() => s.previousValue, throwsA(isA<Exception>()));
-      });
-
       group('autoDispose', () {
         test('check last subscriber disposes', () {
           final List<String> list = <String>['a', 'b', 'c'];
