+++ packages/signals_flutter/pubspec.yaml
@@ -1,6 +1,6 @@
 name: signals_flutter
 description: The signals library exposes four core functions which are the building blocks to model any business logic you can think of.
-version: 6.0.0-rc.1
+version: 6.0.0-rc.2
 repository: https://github.com/rodydavis/signals.dart
 homepage: https://dartsignals.dev/
 documentation: https://dartsignals.dev/
@@ -12,7 +12,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  signals_core: ^6.0.0-rc.1
+  signals_core: ^6.0.0-rc.2
 
 dev_dependencies:
   flutter_test:
